More hot-reloader-turbopack refactors
