+++ packages/next/src/server/dev/turbopack-utils.ts
@@ -34,7 +34,11 @@ import {
   type ClientBuildManifest,
   srcEmptySsgManifest,
 } from '../../build/webpack/plugins/build-manifest-plugin'
-import type { SetupOpts } from '../lib/router-utils/setup-dev-bundler'
+import {
+  propagateServerField,
+  type ServerFields,
+  type SetupOpts,
+} from '../lib/router-utils/setup-dev-bundler'
 import { isInterceptionRouteRewrite } from '../../lib/generate-interception-routes-rewrites'
 import type {
   Issue,
@@ -43,6 +47,7 @@ import type {
   StyledString,
   Endpoint,
   WrittenEndpoint,
+  Entrypoints,
 } from '../../build/swc'
 import {
   MAGIC_IDENTIFIER_REGEX,
@@ -53,6 +58,7 @@ import {
   HMR_ACTIONS_SENT_TO_BROWSER,
   type HMR_ACTION_TYPES,
 } from './hot-reloader-types'
+import * as Log from '../../build/output/log'
 
 export interface InstrumentationDefinition {
   files: string[]
@@ -823,9 +829,9 @@ export async function handleRouteType({
   fontManifests: FontManifests
   loadableManifests: LoadableManifests
   currentEntrypoints: CurrentEntrypoints
-  handleRequireCacheClearing: HandleRequireCacheClearing | undefined
-  changeSubscription: ChangeSubscription | undefined
-  readyIds: ReadyIds
+  handleRequireCacheClearing: HandleRequireCacheClearing | undefined // Dev
+  changeSubscription: ChangeSubscription | undefined // Dev
+  readyIds: ReadyIds | undefined // Dev
   page: string
   route: Route
 }) {
@@ -890,7 +896,7 @@ export async function handleRouteType({
           route.dataEndpoint,
           (pageName) => {
             // Report the next compilation again
-            readyIds.delete(page)
+            readyIds?.delete(page)
             return {
               event: HMR_ACTIONS_SENT_TO_BROWSER.SERVER_ONLY_CHANGES,
               pages: [pageName],
@@ -970,7 +976,7 @@ export async function handleRouteType({
             return
           }
           // Report the next compilation again
-          readyIds.delete(page)
+          readyIds?.delete(page)
           return {
             action: HMR_ACTIONS_SENT_TO_BROWSER.SERVER_COMPONENT_CHANGES,
           }
@@ -1047,4 +1053,230 @@ export async function handleRouteType({
       throw new Error(`unknown route type ${(route as any).type} for ${page}`)
     }
   }
+}
+
+export async function handleEntrypoints({
+  entrypoints,
+  opts,
+  serverFields,
+  distDir,
+  globalEntrypoints,
+  currentEntrypoints,
+  changeSubscriptions,
+  changeSubscription,
+  clearChangeSubscription,
+  sendHmr,
+  startBuilding,
+  handleRequireCacheClearing,
+  prevMiddleware,
+  currentIssues,
+  buildManifests,
+  appBuildManifests,
+  pagesManifests,
+  appPathsManifests,
+  middlewareManifests,
+  actionManifests,
+  fontManifests,
+  loadableManifests,
+}: {
+  entrypoints: TurbopackResult<Entrypoints>
+  opts: SetupOpts
+  serverFields: ServerFields
+  distDir: string
+  globalEntrypoints: GlobalEntrypoints
+  currentEntrypoints: CurrentEntrypoints
+  changeSubscriptions: Map<string, Promise<AsyncIterator<any>>>
+  changeSubscription: ChangeSubscription
+  clearChangeSubscription: (
+    page: string,
+    type: 'server' | 'client'
+  ) => Promise<void>
+  sendHmr: (id: string, payload: HMR_ACTION_TYPES) => void
+  startBuilding: (
+    id: string,
+    requestUrl: string | undefined,
+    forceRebuild: boolean
+  ) => () => void
+  handleRequireCacheClearing: HandleRequireCacheClearing
+  prevMiddleware: boolean | undefined
+  currentIssues: CurrentIssues
+  buildManifests: BuildManifests
+  appBuildManifests: AppBuildManifests
+  pagesManifests: PagesManifests
+  appPathsManifests: AppPathsManifests
+  middlewareManifests: MiddlewareManifests
+  actionManifests: ActionManifests
+  fontManifests: FontManifests
+  loadableManifests: LoadableManifests
+}) {
+  globalEntrypoints.app = entrypoints.pagesAppEndpoint
+  globalEntrypoints.document = entrypoints.pagesDocumentEndpoint
+  globalEntrypoints.error = entrypoints.pagesErrorEndpoint
+
+  currentEntrypoints.clear()
+
+  for (const [pathname, route] of entrypoints.routes) {
+    switch (route.type) {
+      case 'page':
+      case 'page-api':
+      case 'app-page':
+      case 'app-route': {
+        currentEntrypoints.set(pathname, route)
+        break
+      }
+      default:
+        Log.info(`skipping ${pathname} (${route.type})`)
+        break
+    }
+  }
+
+  for (const [pathname, subscriptionPromise] of changeSubscriptions) {
+    if (pathname === '') {
+      // middleware is handled below
+      continue
+    }
+
+    if (!currentEntrypoints.has(pathname)) {
+      const subscription = await subscriptionPromise
+      subscription.return?.()
+      changeSubscriptions.delete(pathname)
+    }
+  }
+
+  const { middleware, instrumentation } = entrypoints
+  // We check for explicit true/false, since it's initialized to
+  // undefined during the first loop (middlewareChanges event is
+  // unnecessary during the first serve)
+  if (prevMiddleware === true && !middleware) {
+    // Went from middleware to no middleware
+    await clearChangeSubscription('middleware', 'server')
+    sendHmr('middleware', {
+      event: HMR_ACTIONS_SENT_TO_BROWSER.MIDDLEWARE_CHANGES,
+    })
+  } else if (prevMiddleware === false && middleware) {
+    // Went from no middleware to middleware
+    sendHmr('middleware', {
+      event: HMR_ACTIONS_SENT_TO_BROWSER.MIDDLEWARE_CHANGES,
+    })
+  }
+
+  if (opts.nextConfig.experimental.instrumentationHook && instrumentation) {
+    const processInstrumentation = async (
+      displayName: string,
+      name: string,
+      prop: 'nodeJs' | 'edge'
+    ) => {
+      const writtenEndpoint = await instrumentation[prop].writeToDisk()
+      handleRequireCacheClearing(displayName, writtenEndpoint)
+      processIssues(currentIssues, name, writtenEndpoint)
+    }
+    await processInstrumentation(
+      'instrumentation (node.js)',
+      'instrumentation.nodeJs',
+      'nodeJs'
+    )
+    await processInstrumentation(
+      'instrumentation (edge)',
+      'instrumentation.edge',
+      'edge'
+    )
+    await loadMiddlewareManifest(
+      distDir,
+      middlewareManifests,
+      'instrumentation',
+      'instrumentation'
+    )
+    await writeManifests({
+      rewrites: opts.fsChecker.rewrites,
+      distDir,
+      buildManifests,
+      appBuildManifests,
+      pagesManifests,
+      appPathsManifests,
+      middlewareManifests,
+      actionManifests,
+      fontManifests,
+      loadableManifests,
+      currentEntrypoints,
+    })
+
+    serverFields.actualInstrumentationHookFile = '/instrumentation'
+    await propagateServerField(
+      opts,
+      'actualInstrumentationHookFile',
+      serverFields.actualInstrumentationHookFile
+    )
+  } else {
+    serverFields.actualInstrumentationHookFile = undefined
+    await propagateServerField(
+      opts,
+      'actualInstrumentationHookFile',
+      serverFields.actualInstrumentationHookFile
+    )
+  }
+  if (middleware) {
+    const processMiddleware = async () => {
+      const writtenEndpoint = await middleware.endpoint.writeToDisk()
+      handleRequireCacheClearing('middleware', writtenEndpoint)
+      processIssues(currentIssues, 'middleware', writtenEndpoint)
+      await loadMiddlewareManifest(
+        distDir,
+        middlewareManifests,
+        'middleware',
+        'middleware'
+      )
+      serverFields.middleware = {
+        match: null as any,
+        page: '/',
+        matchers:
+          middlewareManifests.get('middleware')?.middleware['/'].matchers,
+      }
+    }
+    await processMiddleware()
+
+    changeSubscription(
+      'middleware',
+      'server',
+      false,
+      middleware.endpoint,
+      async () => {
+        const finishBuilding = startBuilding('middleware', undefined, true)
+        await processMiddleware()
+        await propagateServerField(
+          opts,
+          'actualMiddlewareFile',
+          serverFields.actualMiddlewareFile
+        )
+        await propagateServerField(opts, 'middleware', serverFields.middleware)
+        await writeManifests({
+          rewrites: opts.fsChecker.rewrites,
+          distDir,
+          buildManifests,
+          appBuildManifests,
+          pagesManifests,
+          appPathsManifests,
+          middlewareManifests,
+          actionManifests,
+          fontManifests,
+          loadableManifests,
+          currentEntrypoints,
+        })
+
+        finishBuilding()
+        return { event: HMR_ACTIONS_SENT_TO_BROWSER.MIDDLEWARE_CHANGES }
+      }
+    )
+    prevMiddleware = true
+  } else {
+    middlewareManifests.delete('middleware')
+    serverFields.actualMiddlewareFile = undefined
+    serverFields.middleware = undefined
+    prevMiddleware = false
+  }
+  await propagateServerField(
+    opts,
+    'actualMiddlewareFile',
+    serverFields.actualMiddlewareFile
+  )
+  await propagateServerField(opts, 'middleware', serverFields.middleware)
 }
